requirements.txt
django
psycopg2
six
django-leaflet
